+++ src/main/java/com/gym/crm/app/repository/impl/TrainerRepositoryImpl.java
@@ -14,13 +14,13 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 @Repository
 public class TrainerRepositoryImpl implements TrainerRepository {
     private static final Logger logger = LoggerFactory.getLogger(TrainerRepositoryImpl.class);
 
-    private final AtomicInteger trainerCounter = new AtomicInteger(1);
+    private final AtomicLong trainerCounter = new AtomicLong(1);
 
     private Map<String, Trainer> trainerStorage;
 
@@ -43,7 +43,7 @@ public class TrainerRepositoryImpl implements TrainerRepository {
         }
 
         Trainer trainerWithId = trainer.toBuilder()
-                .id((long) trainerCounter.getAndIncrement())
+                .id(trainerCounter.getAndIncrement())
                 .build();
 
         trainerStorage.put(key, trainerWithId);
